Added bullets which the enemy shoots
Bullets can hit a player, hit an obstacle, or go off the map. Each of
these will result in the bullet being deleted, and if it hits the
player, it will adjust the player's health. The player's health goes
down by the damage of the bullet and enters an invulnerable state for 3
seconds after being hit. The player will die if it's health goes below
0, for now it just deletes the player object when this happens.