Added basic test menu with reset button
Added a menu which can be opened using start or escape. This sets time scale to 0 which pauses the game movement, and also switches player input scheme to "UI" which allows them to click on menu options.

Currently, the only button to press is "RESET" which resets the scene and respawns all enemies