Dash is back!
The Ability abstract class is done for now, as in it fully works. I have implement the Dash ability as a child class of Ability and the player is once again able to boost themselves around and, more importantly, know that the boost will end instead of launching them off into the void.

There is still much work to do on the ability class, such as implementing the charge system so we can have multiple dashes that cooldown independently or subsequently.