Jump is now implemented using the new ability system.
Jump uses the Ability class's cooldown to add a short delay before jump can be used again. Number of jumps is now controlled by the ability itself, set initially by PlayerStats. Each jump uses the Ability class UseCharge(), and charges are refreshed when the PlayerGroundedEvent is raised in the MovementController.

A new GameObject called PlayerEventListeners has been added to hold all event listeners needed for the player. When the PlayerGroundedEvent is raised, it uses a GameEventListener component to call any functions added to it. In this case, it calls the RefreshAllCharges() function in the Jump Ability to reset the number of jump available to the player.