Double-tapping the down button will now cause the player to drop through a platform if it is configured as a one way platform.
Platform dropping is part of the Jump ability. There are 2 new functions for starting and stopping a drop, and a coroutine that controls the delay between starting and stopping the "dropping" condition. Right now this is the same as the minimum time between jumping and seems to work fine.

If there are multiple levels of one-way platforms the player will drop through the first and as long as the "isDropping" condition is false by the time they hit the next one, they will land on it. My original goal was to maintain the dropped condition as long as the down button was held after the double tap, but I think that will require custom input processing code that I don't want to get in to yet.

One additional strange interaction is if a player does not jump high enough to land on a one way platform, once their Y velocity is negative (falling) they will resume collision with the platform, potentially halting their X-Axis movement without warning. I couldn't think of a good way to prevent this, so I have left it as is for this version so I can wrap up a few other pieces of this overhaul.