Player no longer gains infinite jumps while wall jumping.
Wall jumping still does not use a jump charge, but performing a normal air jump after wall jumping correctly spends a jump charge. When wall jumping, the player is pushed upwards at their max jump velocity, and in the X direction opposite of the the wall they are sliding on for the same amount of force. We can create a new variable for wall jump X force and tweak that until it feels better, but it is functional for now.